Added Scheme object, in the 'Quasi Conservative Scheme' sense. Have also modified Riemann Solver slightly to be compatible with new StateVector. Still needs more work. Have also removed explicit keyword from constructors where there was more than one parameter.

+++ mesh/mesh.h
@@ -7,7 +7,7 @@ namespace mesh {
   
   class Mesh1D {
     private:
-      size_t N;                          // Number of cells
+      int N;                             // Number of cells
       double width;                      // Width of domain
       std::vector<double> dx;            // Width of cells
       std::vector<flow::StateVector> U;  // Flow value vector (conservative formualation)
@@ -16,7 +16,7 @@ namespace mesh {
       void init();                       // Initialise the cell size vector
 
     public:
-      Mesh1D(size_t _N, const double& _width);
+      Mesh1D(int _N, const double& _width);
       
       StateVector& meshPoint(const int& j);
       StateVector& slopePoint(const int& j);
